pyproject.toml
[project]
name = "cippa-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiogram>=3.13.1",
    "aiogram-dialog>=2.2.0",
    "asyncpg>=0.30.0",
    "environs>=11.0.0",
    "nats-py>=2.9.0",
    "orjson>=3.10.10",
    "pydantic>=2.9.2",
    "pydantic-settings>=2.6.0",
    "sqlalchemy>=2.0.36",
    "uvicorn>=0.32.0",
    "uvloop>=0.21.0",
    "greenlet>=3.1.1",
]
